deployment/requirements.txt
numpy==2.0.2
pandas==2.2.2
pandas-datareader==0.10.0
pandas-gbq==0.28.0
pandas-stubs==2.2.2.240909
seaborn==0.13.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
matplotlib-venn==1.1.2
joblib==1.4.2
scikit-image==0.25.2
scikit-learn==1.6.1
streamlit
xgboost
google.generativeai